Added pages loading

+++ app/src/main/java/com/neliry/banancheg/videonotes/models/Conspectus.kt
@@ -3,5 +3,5 @@ package com.neliry.banancheg.videonotes.models
 class Conspectus(id:String? = null,
                  name:String? = null,
                  var previewUrl: String? = null,
-                 var time:  Int? = null,
+                 var time:  Long? = null,
                  var videoUrl: String? = null): BaseItem(id,name)
